Fixing SEAMSECURITY-132
This was a problem with eclipselink (possibly other JPA impls as well)
where the entity given was the fully qualified name of the class instead
of the actual name of the entity / simple class name.

Also made the JpaIdentityStore class work if you specify a different
entity name besides the simple class name. All JPAQL queries should be
working correctly.